fix tests

+++ core/src/test/java/me/paulbares/query/TestPeriodShiftProcedure.java
@@ -49,21 +49,21 @@ public class TestPeriodShiftProcedure {
       new PeriodComparisonExecutor.ShiftProcedure(
               period,
               Map.of(PeriodUnit.YEAR, refPos[0], PeriodUnit.MONTH, refPos[1]),
-              indexByPeriodUnit).test(point, new Field[]{new Field("year", long.class), new Field("month", int.class)});
+              indexByPeriodUnit).test(point, new Field[]{new Field("year", long.class), new Field("month", long.class)}); // use long and make sure we get long at the end
       return point;
     };
 
-    Assertions.assertThat(f.apply(new Object[]{2022, 1}, new String[]{"y", "q-1"})).containsExactly(2021, 12);
-    Assertions.assertThat(f.apply(new Object[]{2022, 2}, new String[]{"y", "q-1"})).containsExactly(2022, 1);
-    Assertions.assertThat(f.apply(new Object[]{2022, 1}, new String[]{"y-1", "q-1"})).containsExactly(2020, 12);
-    Assertions.assertThat(f.apply(new Object[]{2022, 2}, new String[]{"y-1", "q-1"})).containsExactly(2021, 1);
+    Assertions.assertThat(f.apply(new Object[]{2022, 1}, new String[]{"y", "q-1"})).containsExactly(2021l, 12l);
+    Assertions.assertThat(f.apply(new Object[]{2022, 2}, new String[]{"y", "q-1"})).containsExactly(2022l, 1l);
+    Assertions.assertThat(f.apply(new Object[]{2022, 1}, new String[]{"y-1", "q-1"})).containsExactly(2020l, 12l);
+    Assertions.assertThat(f.apply(new Object[]{2022, 2}, new String[]{"y-1", "q-1"})).containsExactly(2021l, 1l);
 
-    Assertions.assertThat(f.apply(new Object[]{2022, 12}, new String[]{"y", "q+1"})).containsExactly(2023, 1);
-    Assertions.assertThat(f.apply(new Object[]{2022, 2}, new String[]{"y", "q+1"})).containsExactly(2022, 3);
-    Assertions.assertThat(f.apply(new Object[]{2022, 12}, new String[]{"y+1", "q+1"})).containsExactly(2024, 1);
-    Assertions.assertThat(f.apply(new Object[]{2022, 2}, new String[]{"y+1", "q+1"})).containsExactly(2023, 3);
+    Assertions.assertThat(f.apply(new Object[]{2022, 12}, new String[]{"y", "q+1"})).containsExactly(2023l, 1l);
+    Assertions.assertThat(f.apply(new Object[]{2022, 2}, new String[]{"y", "q+1"})).containsExactly(2022l, 3l);
+    Assertions.assertThat(f.apply(new Object[]{2022, 12}, new String[]{"y+1", "q+1"})).containsExactly(2024l, 1l);
+    Assertions.assertThat(f.apply(new Object[]{2022, 2}, new String[]{"y+1", "q+1"})).containsExactly(2023l, 3l);
 
-    Assertions.assertThat(f.apply(new Object[]{2022, 1}, new String[]{"y", "q-2"})).containsExactly(2021, 11);
+    Assertions.assertThat(f.apply(new Object[]{2022, 1}, new String[]{"y", "q-2"})).containsExactly(2021l, 11l);
   }
 
   @Test
